Machine Translated Source=extraction only
This file is the Google machine translation of Konrad Blochinger's bio on Jost von Silenen. It is not designed for legibility, but  shows were the superimposed overlaps of the OCR translation fit. I have attached the extracted English translations in .odt, .doc, and .pdf format unformatted-for expediency, it is not a document I have time right now to properly clean up the extraneous characters such as "• «•" and "**• '" There are several other documents that I will be translating from German into French, so that some of the experts in French can compare to von Silenen and the unidentified author.