Add files via upload
Most pdf versions of The Elegant Enigma have faded and high contrast text, rendering poor readability and eye strain due to bright backlight required to read the small font. This is a useful supplement, although missing are the illustrations and contextual type, thus original scans of the PDF (on Internet Archive and NSA/.MIL are essential for comparison). Fine for portable and casual reading as well.